// src/types/api/service/models/NewsAlertScheduleDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * MediaBee Service API
 * MediaBee Service API
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ScheduleRow } from './ScheduleRow';
import {
    ScheduleRowFromJSON,
    ScheduleRowFromJSONTyped,
    ScheduleRowToJSON,
} from './ScheduleRow';

/**
 * 
 * @export
 * @interface NewsAlertScheduleDto
 */
export interface NewsAlertScheduleDto {
    /**
     * 스케줄 ID
     * @type {number}
     * @memberof NewsAlertScheduleDto
     */
    scheduleId?: number;
    /**
     * 알리미 ID
     * @type {number}
     * @memberof NewsAlertScheduleDto
     */
    alertId: number;
    /**
     * 모니터링 ID
     * @type {number}
     * @memberof NewsAlertScheduleDto
     */
    newsSrchId: number;
    /**
     * 스케줄 리스트
     * @type {Array<ScheduleRow>}
     * @memberof NewsAlertScheduleDto
     */
    schedules: Array<ScheduleRow>;
    /**
     * 종료 일시
     * @type {Date}
     * @memberof NewsAlertScheduleDto
     */
    expireAt?: Date;
}

/**
 * Check if a given object implements the NewsAlertScheduleDto interface.
 */
export function instanceOfNewsAlertScheduleDto(value: object): value is NewsAlertScheduleDto {
    if (!('alertId' in value) || value['alertId'] === undefined) return false;
    if (!('newsSrchId' in value) || value['newsSrchId'] === undefined) return false;
    if (!('schedules' in value) || value['schedules'] === undefined) return false;
    return true;
}

export function NewsAlertScheduleDtoFromJSON(json: any): NewsAlertScheduleDto {
    return NewsAlertScheduleDtoFromJSONTyped(json, false);
}

export function NewsAlertScheduleDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): NewsAlertScheduleDto {
    if (json == null) {
        return json;
    }
    return {
        
        'scheduleId': json['scheduleId'] == null ? undefined : json['scheduleId'],
        'alertId': json['alertId'],
        'newsSrchId': json['newsSrchId'],
        'schedules': ((json['schedules'] as Array<any>).map(ScheduleRowFromJSON)),
        'expireAt': json['expireAt'] == null ? undefined : (new Date(json['expireAt'])),
    };
}

export function NewsAlertScheduleDtoToJSON(value?: NewsAlertScheduleDto | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'scheduleId': value['scheduleId'],
        'alertId': value['alertId'],
        'newsSrchId': value['newsSrchId'],
        'schedules': ((value['schedules'] as Array<any>).map(ScheduleRowToJSON)),
        'expireAt': value['expireAt'] == null ? undefined : ((value['expireAt']).toISOString()),
    };
}