// src/types/api/service/models/ScheduleRow.ts
/* tslint:disable */
/* eslint-disable */
/**
 * MediaBee Service API
 * MediaBee Service API
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 스케줄 리스트
 * @export
 * @interface ScheduleRow
 */
export interface ScheduleRow {
    /**
     * 월요일
     * @type {boolean}
     * @memberof ScheduleRow
     */
    isMonday: boolean;
    /**
     * 화요일
     * @type {boolean}
     * @memberof ScheduleRow
     */
    isTuesday: boolean;
    /**
     * 수요일
     * @type {boolean}
     * @memberof ScheduleRow
     */
    isWednesday: boolean;
    /**
     * 목요일
     * @type {boolean}
     * @memberof ScheduleRow
     */
    isThursday: boolean;
    /**
     * 금요일
     * @type {boolean}
     * @memberof ScheduleRow
     */
    isFriday: boolean;
    /**
     * 토요일
     * @type {boolean}
     * @memberof ScheduleRow
     */
    isSaturday: boolean;
    /**
     * 일요일
     * @type {boolean}
     * @memberof ScheduleRow
     */
    isSunday: boolean;
    /**
     * 발송 시간 (0-24)
     * @type {number}
     * @memberof ScheduleRow
     */
    hour: number;
    /**
     * 발송 분 (0-59)
     * @type {number}
     * @memberof ScheduleRow
     */
    minute: number;
}

/**
 * Check if a given object implements the ScheduleRow interface.
 */
export function instanceOfScheduleRow(value: object): value is ScheduleRow {
    if (!('isMonday' in value) || value['isMonday'] === undefined) return false;
    if (!('isTuesday' in value) || value['isTuesday'] === undefined) return false;
    if (!('isWednesday' in value) || value['isWednesday'] === undefined) return false;
    if (!('isThursday' in value) || value['isThursday'] === undefined) return false;
    if (!('isFriday' in value) || value['isFriday'] === undefined) return false;
    if (!('isSaturday' in value) || value['isSaturday'] === undefined) return false;
    if (!('isSunday' in value) || value['isSunday'] === undefined) return false;
    if (!('hour' in value) || value['hour'] === undefined) return false;
    if (!('minute' in value) || value['minute'] === undefined) return false;
    return true;
}

export function ScheduleRowFromJSON(json: any): ScheduleRow {
    return ScheduleRowFromJSONTyped(json, false);
}

export function ScheduleRowFromJSONTyped(json: any, ignoreDiscriminator: boolean): ScheduleRow {
    if (json == null) {
        return json;
    }
    return {
        
        'isMonday': json['isMonday'],
        'isTuesday': json['isTuesday'],
        'isWednesday': json['isWednesday'],
        'isThursday': json['isThursday'],
        'isFriday': json['isFriday'],
        'isSaturday': json['isSaturday'],
        'isSunday': json['isSunday'],
        'hour': json['hour'],
        'minute': json['minute'],
    };
}

export function ScheduleRowToJSON(value?: ScheduleRow | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'isMonday': value['isMonday'],
        'isTuesday': value['isTuesday'],
        'isWednesday': value['isWednesday'],
        'isThursday': value['isThursday'],
        'isFriday': value['isFriday'],
        'isSaturday': value['isSaturday'],
        'isSunday': value['isSunday'],
        'hour': value['hour'],
        'minute': value['minute'],
    };
}